refactor: conftest mock nodes
