+++ tests/unittests/test_nodes/test_host_nodes/test_img_detections_bridge_node.py
@@ -9,7 +9,7 @@ from pytest import FixtureRequest
 from depthai_nodes import ImgDetectionExtended, ImgDetectionsExtended
 from depthai_nodes.node import ImgDetectionsBridge
 
-from .conftest import Output
+from tests.utils import OutputMock
 from .utils.create_message import (
     create_img_detections,
     create_img_detections_extended,
@@ -41,7 +41,7 @@ def test_initialization(bridge: ImgDetectionsBridge):
 
 
 def test_building(bridge: ImgDetectionsBridge):
-    bridge.build(Output())
+    bridge.build(OutputMock())
 
 
 @pytest.mark.parametrize(
@@ -58,7 +58,7 @@ def test_processing(
         img_detections_type
     )
 
-    o_dets = Output()
+    o_dets = OutputMock()
     bridge.build(o_dets, ignore_angle=True)
     q_dets = o_dets.createOutputQueue()
     q_dets_transformed = bridge.out.createOutputQueue()
