+++ tests/unittests/test_nodes/test_host_nodes/test_tiling_node.py
@@ -6,7 +6,7 @@ import pytest
 
 from depthai_nodes.node import Tiling
 
-from .conftest import Output
+from tests.utils import OutputMock
 
 
 @pytest.fixture
@@ -32,7 +32,7 @@ def test_initialization(tiling):
 
 
 def test_build(tiling):
-    output = Output()
+    output = OutputMock()
     grid_size = (2, 2)
     img_shape = (1280, 720)
     nn_shape = np.array([300, 300])
@@ -68,7 +68,7 @@ def test_build(tiling):
 
 
 def test_build_global_detection(tiling, dummy_img_frame):
-    output = Output()
+    output = OutputMock()
     grid_size = (2, 2)
     img_shape = (dummy_img_frame.getWidth(), dummy_img_frame.getHeight())
     nn_shape = np.array([300, 300])
@@ -93,7 +93,7 @@ def test_build_global_detection(tiling, dummy_img_frame):
 
 
 def test_process(tiling, dummy_img_frame, duration):
-    output = Output()
+    output = OutputMock()
     grid_size = (2, 2)
     img_shape = (dummy_img_frame.getWidth(), dummy_img_frame.getHeight())
     nn_shape = np.array([300, 300])
