+++ tests/unittests/test_nodes/test_threaded_host_nodes/test_host_parsing_neural_network_node.py
@@ -1,8 +1,8 @@
 import depthai as dai
 import pytest
-from stability_tests.conftest import Input, NeuralNetworkMock, PipelineMock
 
 from depthai_nodes.node import HostParsingNeuralNetwork, YOLOExtendedParser
+from tests.utils import InputMock, NeuralNetworkMock, PipelineMock
 
 
 @pytest.fixture
@@ -20,7 +20,7 @@ def test_yolo(pipeline: PipelineMock):
         dai.getModelFromZoo(dai.NNModelDescription("luxonis/yolov6-nano", "RVC2"))
     )
     nn = pipeline.create(HostParsingNeuralNetwork).build(
-        input=Input(), nn_source=nn_archive, fps=30
+        input=InputMock(), nn_source=nn_archive, fps=30
     )
     parser = nn.getParser()
     assert isinstance(parser, YOLOExtendedParser)
@@ -34,4 +34,4 @@ def test_unsupported(pipeline: PipelineMock):
     )
     nn = pipeline.create(HostParsingNeuralNetwork)
     with pytest.raises(ValueError):
-        nn.build(input=Input(), nn_source=nn_archive, fps=30)
+        nn.build(input=InputMock(), nn_source=nn_archive, fps=30)
